requirements.txt
streamlit
openai
openai-whisper
python-dotenv
sentence-transformers
transformers
torch
ffmpeg-python
pydub
yt-dlp
tqdm
numpy
scikit-learn

